backend/requirements.txt
fastapi
uvicorn
pydantic-settings
python-multipart
openai
python-dotenv
langchain
tiktoken
faiss-cpu
pypdf
sqlalchemy
psycopg2-binary
